GoogleSheets.jl Documentation

GoogleSheets.jl Documentation

Types

Functions

Index

Types

GoogleSheets.GoogleSheetsClient — Type

A Google Sheets client.

GoogleSheets.Spreadsheet — Type

A spreadsheet.

GoogleSheets.Sheet — Type

A sheet in a spreadsheet.

GoogleSheets.CellRange — Type

A range of cells within a spreadsheet.

GoogleSheets.CellRanges — Type

Multiple ranges of cells within a spreadsheet.

GoogleSheets.CellRangeValues — Type

A range of cell values within a spreadsheet.

GoogleSheets.UpdateSummary — Type

Summary of updated updated cells.

GoogleSheets.CellIndexRange1D — Type

1D Range of cell values.

GoogleSheets.CellIndexRange2D — Type

2D Range of cell values.

GoogleSheets.CellFormat — Type

Formatting for a cell.

DataFrames.DataFrame — Type

Creates a DataFrame from spreadsheet range values. The first row is converted to column names. All other rows are converted to string values.

Functions

GoogleSheets.sheets_client — Function

Creates a client for accessing Google Sheets.

GoogleSheets.sheet_names — Function

Gets the names of the sheets in the spreadsheet.

GoogleSheets.sheets — Function

Gets the sheets in the spreadsheet.

GoogleSheets.batch_update! — Function

Applies one or more updates to a spreadsheet.

Each request is validated before being applied. If any request is not valid then the entire request will fail and nothing will be applied.

Common batch_update! functionality: Charts: https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/charts Filters: https://developers.google.com/sheets/api/guides/filters Basic formatting: https://developers.google.com/sheets/api/samples/formatting Conditional formatting: https://developers.google.com/sheets/api/samples/conditional-formatting Conditional formatting: https://developers.google.com/sheets/api/guides/conditional-format

GoogleSheets.add_sheet! — Function

Adds a new sheet to a spreadsheet.

GoogleSheets.delete_sheet! — Function

Removes a sheet from a spreadsheet.

Removes a sheet from a spreadsheet.

GoogleSheets.freeze! — Function

Freeze rows and columns in a sheet.

Base.append! — Function

Append rows and columns to a sheet.

GoogleSheets.insert_rows! — Function

Insert rows into to a sheet.

GoogleSheets.insert_cols! — Function

Insert columns into to a sheet.

GoogleSheets.delete_rows! — Function

Delete rows from a sheet.

GoogleSheets.delete_cols! — Function

Delete columns from a sheet.

GoogleSheets.meta — Function

Gets metadata about a spreadsheet.

Gets metadata about a spreadsheet key-value pair.

Gets metadata about a spreadsheet sheet.

Gets metadata about a spreadsheet sheet.

Gets metadata about a spreadsheet sheet.

Base.show — Function

Prints metadata about a spreadsheet.

Prints metadata about a spreadsheet sheet.

Prints metadata about a spreadsheet sheet.

Prints metadata about a spreadsheet sheet.

Base.get — Function

Gets a range of cell values from a spreadsheet.

Gets multiple ranges of cell values from a spreadsheet.

GoogleSheets.update! — Function

Updates a range of cell values in a spreadsheet.

Arguments

client::GoogleSheetsClient: client for interacting with Google Sheets.

range::CellRange: cell range to modify.

values::Array{<:Any,2}: values to place in the spreadsheet.

raw::Bool=false: true treats values as raw, unparsed values and and are simply inserted as a string. false treats values exactly as if they were entered into the Google Sheets UI, for example "=A1+B1" is a formula.

Updates a range of cell values in a spreadsheet.

Arguments

client::GoogleSheetsClient: client for interacting with Google Sheets.

range::CellRange: cell range to modify.

df::DataFrame: dataframe of values.

kwargs...: keyword arguments.

GoogleSheets.clear! — Function

Clears a range of cell values in a spreadsheet.

GoogleSheets.format! — Function

Formats a range of cells.

GoogleSheets.format_conditional! — Function

Formats a range of cells if a condition is met.

GoogleSheets.format_color_gradient! — Function

Sets color gradient formatting.

GoogleSheets.update_default_rate_limiter — Function

Update the default rate limiter.

Update the default rate limiter.

GoogleSheets.credentials_file — Function

Gets the credentials file needed to log into Google. The file is loaded from the GOOGLESHEETSCREDENTIALS environment variable if it is present; otherwise, it is loaded from the configuration directory, which defaults to ~/.julia/config/google_sheets/.

See the python quick start reference for a link to generate credentials. https://developers.google.com/sheets/api/quickstart/python

See the API credentials page to create or revoke credentials. https://console.developers.google.com/apis/credentials

Index

DataFrames.DataFrame

GoogleSheets.CellFormat

GoogleSheets.CellIndexRange1D

GoogleSheets.CellIndexRange2D

GoogleSheets.CellRange

GoogleSheets.CellRangeValues

GoogleSheets.CellRanges

GoogleSheets.GoogleSheetsClient

GoogleSheets.Sheet

GoogleSheets.Spreadsheet

GoogleSheets.UpdateSummary

Base.append!

Base.get

Base.show

GoogleSheets.add_sheet!

GoogleSheets.batch_update!

GoogleSheets.clear!

GoogleSheets.credentials_file

GoogleSheets.delete_cols!

GoogleSheets.delete_rows!

GoogleSheets.delete_sheet!

GoogleSheets.format!

GoogleSheets.format_color_gradient!

GoogleSheets.format_conditional!

GoogleSheets.freeze!

GoogleSheets.insert_cols!

GoogleSheets.insert_rows!

GoogleSheets.meta

GoogleSheets.sheet_names

GoogleSheets.sheets

GoogleSheets.sheets_client

GoogleSheets.update!

GoogleSheets.update_default_rate_limiter